--- script/README
@@ -1,34 +1,0 @@
-Создайте три новых класса:
--Card Это класс, создающий карточку. 
-
-    Добавьте ему методы constructor, like и remove.  +
-    И ещё один — create. Он будет создавать DOM-элемент карточки. +
-
--CardList Это класс для хранения и отрисовки карточек. 
-
-    Метод constructor этого класса должен принимать два параметра:
-        DOM-элемент — контейнер, куда нужно складывать карточки;
-        массив карточек, которые будут на странице при загрузке.
-    Ещё у класса CardList должны быть два метода:
-        addCard для добавления карточки в список;
-        render для отрисовки карточек при загрузке страницы.
-
--Popup Это класс для всплывающего окна. 
-
-    Добавьте ему методы open и close, чтобы показывать и скрывать попап. 
-    Есть два подхода, как можно реализовать всплывающие окна:
-        сделать единый контейнер для всех попапов и менять его содержимое при открытии;
-        сделать независимые попапы в разных контейнерах.
-        //Первый способ одновременно лучше и сложнее. Но вы сами можете выбрать, как реализовать попап.
-
-Общие комментарии
-Эта самостоятельная работа отличается от предыдущих. 
-До сих пор от вас требовалось реализовать какой-то функционал. 
-Вы думали, как заставить кнопки работать или сделать так, чтобы попап появлялся и исчезал. 
-Сейчас же весь функционал уже готов. 
-Ваша задача — правильно организовать код.
-
-Главное требование — классы Card, CardList и Popup должны присутствовать в коде. 
-Но организация кода — задача творческая. 
-Поэтому кроме перечисленных классов вы можете придумать свои, если посчитаете нужным. 
-На код-ревью мы обязательно прокомментируем, насколько ваше решение изящно.
