+++ src/blocks/place-card/place-card.css
@@ -1,5 +1,4 @@
 @import url(./__image/place-card__image.css);
-@import url(./__image/_city/_karachaevsk/place-card__image_city_karachaevsk.css);
 @import url(./__delete-icon/place-card__delete-icon.css);
 @import url(./__description/place-card__description.css);
 @import url(./__name/place-card__name.css);
