+++ src/blocks/popup/__span-error/popup__span-error.css
