+++ src/script/.jshintrc
