+++ src/script/index.js
@@ -1,199 +1,7 @@
-
-/* Классы */
-
-class Card {
-
-  constructor(title, url, idCard) {
-    this.cardElement = this.create(title,url, idCard);
-    this.idCard = idCard;
-  }
-  
-  remove(event) {
-    event.target.closest('.place-card').parentElement.removeChild(this.cardElement);
-  }
-
-  like(event) {
-    event.target.classList.toggle('place-card__like-icon_liked');
-  }
-
-  create(title, url, idCard) {
-
-    const cardContainer = document.createElement('div');
-    const cardImageContainer = document.createElement('div');
-    const cardButtonDelete = document.createElement('button');
-    const cardDescriptionContainer = document.createElement('div');
-    const cardTitleElement = document.createElement('h3');
-    const cardButtonLike = document.createElement('button');
-
-    cardContainer.classList.add('place-card');
-    cardImageContainer.classList.add('place-card__image');
-    cardImageContainer.style.backgroundImage = `url(${url})`;
-    cardButtonDelete.classList.add('place-card__delete-icon');
-    cardDescriptionContainer.classList.add('place-card__description');
-    cardTitleElement.classList.add('place-card__name');
-    cardTitleElement.textContent = title;
-    cardButtonLike.classList.add('place-card__like-icon');
-
-    if(idCard === api.user._id){
-      cardButtonDelete.classList.add('place-card__delete-icon_display-block');
-    }
-
-    cardImageContainer.appendChild(cardButtonDelete);
-    cardDescriptionContainer.appendChild(cardTitleElement);
-    cardDescriptionContainer.appendChild(cardButtonLike);
-    cardContainer.appendChild(cardImageContainer);
-    cardContainer.appendChild(cardDescriptionContainer);
-
-
-    cardButtonLike.addEventListener('click', this.like);
-    cardButtonDelete.addEventListener('click', (event) => {this.remove(event);});
-    return cardContainer;
-  }
-
-}
-class CardList {
-  constructor(container, /*listCards,*/ popup) {
-    this.container  = container;
-    // this.listCards  = listCards;
-    this.popupImage = new Popup(popup);
-    
-    this.container.addEventListener('click', this.openImagePopup.bind(this));
-    // this.render();
-  }
-
-  addCard(title, url, idCard) {
-    const { cardElement } = new Card(title, url, idCard);
-    this.container.appendChild(cardElement);
-  }
-
-  // render() {
-  //   this.listCards.forEach(element => {
-  //     this.addCard(element.name, element.link);
-  //   });
-  // }
-
-  openImagePopup(event){
-    if(event.target.classList.contains('place-card__image')){
-      this.popupImage.openImage(event.target.style.backgroundImage.slice(5,-2),this.popupImage);
-    }
-  }
-
-}
-
-class Popup {
-  constructor(popup, buttonOpenPopup) {
-    
-    this.popup              = popup;
-    this.contentPopup       = this.popup.querySelector('.popup__content');
-    this.buttonClose        = this.popup.querySelector('.popup__close');
-    this.img                = document.createElement('img');
-
-    if(!!buttonOpenPopup){
-      this.buttonOpenPopup  = buttonOpenPopup;
-      this.buttonOpenPopup.addEventListener('click', event => {this.open(event);});
-    }    
-
-    if(!this.contentPopup.classList.contains('.popup__img')){
-      this.form = this.contentPopup.querySelector('.popup__form');
-    }
-
-    this.buttonClose.addEventListener('click', event => {this.close(event);});
-  }
-  
-  open() {
-    this.popup.classList.add('popup_is-opened');
-  }
-
-  openImage(url, popup){
-    this.img.src = url;
-    this.img.classList.add('popup__img');
-    this.contentPopup.appendChild(this.img);
-    popup.open();
-  }
-
-  close() {
-    if(this.contentPopup.classList.contains('popup__img')){
-      this.contentPopup.removeChild(this.img);
-    } else {
-      this.form.reset();
-    }
-    this.popup.classList.remove('popup_is-opened');
-  }
-
-}
-
-class Api {
-  constructor(options) {
-    // тело конструктора
-    this.url      = options.baseUrl;
-    this.headers  = options.headers;
-    this.user     = this.getUserInfo(); 
-  }
-
-  // Формируем запрос на сервер
-  sendRequest(url, method, body) {
-    return fetch(`${this.url}${url}`, {
-			method: method.toUpperCase(), 
-			headers: this.headers,
-      body: JSON.stringify(body)
-		})
-    .then(res => {
-        if(res.ok) {
-            return res.json();
-        }
-        return Promise.reject(`Ошибка: ${res.status}`);
-    })
-    .catch(err => Promise.reject(err));
-  }
-
-  // загружаем с сервера начальный набор карточек
-  getInitialCards() {
-    this.sendRequest('/cards','get')
-    // .then(res => res.filter(item => item.owner._id == this.user._id)) // Удалить строку для отображения всех карточек
-    .then(res => {
-      res.forEach(item => {
-        cardList.addCard(item.name, item.link, item.owner._id);
-      });
-    })
-    .catch(err => console.log(err));
-  }
-
-  // получаем информацию о пользователе
-  getUserInfo() {
-    this.sendRequest('/users/me', 'get').then(res => {
-      this.user = res;
-      userInfoName.textContent          = res.name;
-      userInfoJob.textContent           = res.about;
-      userAvatar.style.backgroundImage  = `url(${res.avatar})`;
-    }).catch(err => console.log(err));
-  }
-
-  // добавляем карточку
-  addNewCard(body) {
-    return this.sendRequest('/cards', 'POST', body);  
-  }
-  
-  // отправляем информацию о пользователе
-  setUserInfo(body) {  
-    return this.sendRequest('/users/me', 'PATCH', body);
-  }
-
-  setUserAvatar(body) {
-    return this.sendRequest('/users/me/avatar','PATCH', body/*{avatar: ''}*/);
-  }
-
-  deleteCard() {
-
-  }
-
-  setLikeCard() {
-    return this.sendRequest('/cards/like/cardId', 'PUT', body);
-  }
-
-  
-  // другие методы работы с API
-}
-
+import "../pages/index.css";
+import {CardList} from './Cardlist.js';
+import {Popup} from './Popup.js';
+import {Api} from './Api.js';
 
 /* Функции */
 
@@ -405,8 +213,3 @@ formEditInfo.elements.inputSecond.addEventListener('input', handleValidate);
 formAddNewCard.elements.inputSecond.addEventListener('input', handleValidate);
 formEditInfo.elements.inputFirst.addEventListener('input', handleValidate);
 formAvatar.elements.inputFirst.addEventListener('input', handleValidate);
-
-/*
-  Отлично, теперь все реализовано правильно
-
-*/
