+++ webpack.config.js
@@ -1,0 +1,54 @@
+const path = require('path');
+const MiniCssExtractPlugin = require("mini-css-extract-plugin");
+const HtmlWebpackPlugin = require('html-webpack-plugin');
+const WebpackMd5Hash = require('webpack-md5-hash');
+const ImageWebpackLoader= require('image-webpack-loader');
+
+module.exports = {
+        entry: { main: './src/script/index.js' },
+        output: {
+            path: path.resolve(__dirname, 'dist'),
+            filename: '[name].[chunkhash].js'
+        },
+        module: {
+            rules: [{
+                test: /\.js$/,
+                use: { loader: "babel-loader" },
+                exclude: /node_modules/
+            },
+            {
+                test: /\.css$/, 
+                use: [MiniCssExtractPlugin.loader, 'css-loader', 'postcss-loader']
+            },
+            {
+                test: /\.(png|jpg|gif|ico|svg)$/,
+                use: [
+                    'file-loader?name=./images/[name].[ext]',
+                    {
+                        loader: 'image-webpack-loader',
+                        options: { 
+                            bypassOnDebug:  true,  
+                            disable:  true,
+                        }
+                    },
+                ],
+            },
+            {
+                test: /\.(eot|ttf|woff|woff2)$/,
+                loader: 'file-loader?name=./vendor/[name].[ext]'
+            }
+        ]
+        },
+        plugins: [ 
+            new MiniCssExtractPlugin({
+                filename: 'style.[contenthash].css'
+            }),
+            new HtmlWebpackPlugin({
+                inject: false,
+                hash: true,
+                template: './src/index.html',
+                filename: 'index.html'
+              }),
+              new WebpackMd5Hash(),
+        ]
+};
